Rename module `compote` -> `compy`.
It matches other names in this project, like `typey` and `loggy`.
