+++ src/compy/README.md
@@ -1,5 +1,5 @@
-@this/compote
-=============
+@this/compy
+===========
 
 This module implements the low level "component system" upon which the rest of
 the framework is built.
