+++ src/compy/export/BaseComponent.js
