+++ src/compy/export/BaseConfig.js
