+++ src/compy/export/IntfComponent.js
