+++ src/compy/export/ThreadlikeComponent.js
