+++ src/compy/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@this/compote",
+  "name": "@this/compy",
   "version": "0.0.1",
   "type": "module",
   "private": true,
