+++ src/webapp-core/export/BaseApplication.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { BaseComponent, BaseConfig } from '@this/compote';
+import { BaseComponent, BaseConfig } from '@this/compy';
 import { DispatchInfo, FullResponse, IncomingRequest, IntfRequestHandler,
   StatusResponse, TypeOutgoingResponse }
   from '@this/net-util';
@@ -114,7 +114,7 @@ export class BaseApplication extends BaseComponent {
    *
    * This class mostly exists to be an easy target to use when subclasses want
    * to define configuration classes in the usual way, without having to
-   * remember the persnickety detail of which class in the `compote` module is
+   * remember the persnickety detail of which class in the `compy` module is
    * the most appropriate one to derive from.
    */
   static Config = class Config extends BaseConfig {
