+++ src/webapp-core/export/ComponentManager.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { BaseComponent, BaseConfig } from '@this/compote';
+import { BaseComponent, BaseConfig } from '@this/compy';
 import { MustBe } from '@this/typey';
