+++ src/webapp-core/private/ServiceUseConfig.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { BaseConfig, Names } from '@this/compote';
+import { BaseConfig, Names } from '@this/compy';
 
 
 /**
