+++ src/webapp-util/export/BaseFileService.js
@@ -4,7 +4,7 @@
 import * as fs from 'node:fs/promises';
 
 import { WallClock } from '@this/clocks';
-import { BaseConfig } from '@this/compote';
+import { BaseConfig } from '@this/compy';
 import { Duration } from '@this/data-values';
 import { Paths, Statter } from '@this/fs-util';
 import { MustBe } from '@this/typey';
